Requeue projects pending deletion
There have been several bugs in the project deletion service and worker.
Resulting in projects stuck in pending delete state, which limits users
to create projects with the same name, keeps stale records in the
database, and all kinds of other trouble.

This post deployment migration requeues all these projects for deletion,
in the hope that most of these could be removed by the updated code.