Detect archive format from the file extension
The entire file name including the extension is used to detect the
format of the archive, like `archive.tar.gz` but this doesn't permit
the addition of more flexible routes that accept names like
`project-name-1.2.tar.gz`.

Only the extension of the file should be used to detect the archive
format so that more flexible archive routes are possible.