Perform LDAP group sync on sign in only for new users
Syncing a user's groups from LDAP every time they sign in,
including on Git operations, is causing performance issues for
some customers. The benefit of sync on sign in is primarily
for new users so this change restricts it as such. After initial
sign in LDAP group sync will be handled by the hourly (default)
sync and the on-demand group membership sync. Fixes #7352