Remove Git circuit breaker
Was introduced in the time that GitLab still used NFS, which is not
required anymore in most cases. By removing this, the API it calls will
return empty responses. This interface has to be removed in the next
major release, expected to be 12.0.

CE MR: https://gitlab.com/gitlab-org/gitlab-ce/merge_requests/22212

Conflicts:
	app/controllers/admin/health_check_controller.rb
	app/controllers/application_controller.rb
	app/helpers/application_settings_helper.rb
	db/schema.rb
	doc/administration/monitoring/prometheus/gitlab_metrics.md
	doc/api/settings.md
	locale/gitlab.pot