EE: Fix EE specific code affected by refactor
- EE: Restore metrics json endpoint for cluster

Use new controller so that we have one less EE prepend.

- EE: Fix assets and views for ClustersController

We moved from Projects::ClustersController to ClustersController.
So, some of the EE only views and assets needs to be updated.

Also update some EE only controller specs to ClustersController

- EE: Update override of can_create_cluster?

The method signature changed in CE to accomodate group as a
`clusterable`. Correspondingly, update the method to use `clusterable`.

This presumes multiple group level clusters is available at the same
license as projet level clusters which is GitLab Premium.

Also fix related specs