Separate SSH host keys from ProjectImportData
As we're using SSH host key detection for both pull and push mirroring,
we can no longer just look the previous state of the host keys up in
the database.

Instead, the frontend can pass the value it currently knows about to
the controller, which can respond with the information we're interested
in - whether the keys it has detected differ from the keys we currently
know about.