Add ApprovalMergeRequestRule and ApprovalProjectRule
These represents rules associated with MRs and projects.
Rules are further associated with User and Group as members.

ApprovalMergeRequestRule is associated with Approvals,
as a record of whom approved what when MR is merged.
This is because recomputation result can change in
the future when project approvers/group members changes.