Fix issue where deleting a remote mirror doesnt remove it from database
Remote mirrors were simply being disabled when the user clicks
the UI delete button. This commit adds a remote mirrors delete
endpoint that correctly destroys the database record.
It is important to correctly delete remote mirrors when the user
requests it as their credentials are present in
`remote_mirrors_attributes`.