Make Project's only_mirror_protected_branches value false by default
This `default_value_for` entry didn't actually do anything because by
default, the UI initialized the flag to `false` and the DB allows `nil`
values.  However, this ensures the model is consistent with that
behavior and avoids the pitfalls we saw in
https://gitlab.com/gitlab-org/gitlab-ce/issues/53515.

This is similar to the change made in
https://gitlab.com/gitlab-org/gitlab-ce/merge_requests/23410 for
remote mirrors.