Allow matching only the repo-root for CODEOWNERS
When wanting to match files in the repo root, but only on the repo's
root. Users would supply `/*` as the pattern.

In our old implementation we would match this pattern as `*`, but
allowing `*` to match directory separators.

In this we treat the `/*` pattern as an entirely separate case:
Matching all files in the root, but only one level deep.