Group SAML SSO page warns when linking account
Linking an account allows the external service to sign that user into
GitLab. We need to make users aware of this to reduce the risk of
phishing attacks where users are maliciously tricked into linking
accounts.

We do this by showing a warning and including the group path and the
URL that the user will be redirected to. We also limit the URL to
ASCII characters to avoid homoglyph attacks.