Support both integer and string as epic count key
Epic state uses enum for saving state (other issuables use
string), because of this inconsistency we override `count_key`
which should return state as string converted to symbol
(.e.g. :closed). In rails 5 these integer states are automatically
converted to strings by Active Record.

But if subquery is used (to optimize searching), Active Record doesn't
convert state into string automatically (probably it can not deal with
the more complex query).

This means that epic state can be either integer (if subquery is used)
or string (if AR converted state to string automatically), so in
`count_key` we should convert state into string only if it's an integer.