Repos leave object pool when moving to a new shard
When a project's repositories are moved to a new shard, the pool
repository is still located at the old shard. This doesn't work anymore,
as git doesn't have access to that.

The unlink RPC is called as it cleans up the remote. The alternate file
is removed too, which doesn't exist, but this doesn't fail the move.

Fixes: https://gitlab.com/gitlab-org/gitlab-ee/issues/9094