Fix Error 500 when deleting a pipeline via the API
GitLab EE has an admin audit log that attempts to call `full_path` on
the entity for metadata. Unlike a project or a group, a CI pipeline
doesn't have this; the best we could do is delegate the project path.
For now, set the value to empty but fill in details of the destroyed
pipeline.

Closes https://gitlab.com/gitlab-org/gitlab-ce/issues/55481