Fix group permissions during LDAP tests
Create the sandbox group with appropriate permissions before the LDAP
group sync test starts so that in subsequent tests the LDAP user has
permission to the group.

This also restores the original (most likely nil) env personal access
token, so that subsequent tests don't perform API calls as the root
user while logged in and performing other actions as a non-root LDAP
user.

Without this change groups would be created by the root user, so when
the LDAP user tried to push, it would not be permitted.