Make it possible to define downstream bridge variables
It is now possible to pass variables to a downstream pipeline from a
bridge job. Variables defined as YAML variables for this bridge job are
going to be populated downstream, and are going to be available in every
job created within a downstream pipeline. These variables are then going
to be passed to GitLab Runner as environment variables.