Perform GeoRepositoryDestroyWorker even when repository does not exist
There is a possibility the project does not have a repository or wiki
on disk, but it can have an entry in the project_registry table, e.g.,
when all attempts to sync the have failed. If we don't perform the
destroy worker it can lead to orphaned entries in the project_registry
table.