Handle repository and wiki in Geo::ProjectSyncWorker independently
Now when we need to initiate repository resync we spawn
ProjectSyncWorker where the need for the sync is checked again
just because we don't know for what exactly purpose this worker
has been spawn so we check the status of both: wiki and repo.
This is suboptimal. We also need this change to be flexible if we need
to spawn some sync that is schedule in the future.