Wrap the approvals required in a fallback rule
This allows us to treat the values from
`Project#approvals_before_merge` and
`MergeRequest#approvals_before_merge` the same way as other approval
rules.

It also makes sure the minimum approvals required in this case, is
equal than or more what was defined on a single rule in the project if
there was a single rule.

If there was no single rule, we use the `approvals_before_merge`
attribute of `Project`