Adds X-Request-Id response header for Workhorse
This change configures workhorse to set the Correlation-ID value in
the response header. This is a common technique which allows errors to
be traced more easily.

This can be done through several methods: 1) users can review their
browser headers and report the value 2) nginx error pages could be
configured to report the correlation-id on the error page, allowing
users to easily report it.

Once the correlation-id has been reported, support engineers or
production staff can use this to further investigate issues using
logging infrastructure, sentry or distributed tracing.