Move related issues components for reusability
We are building related MRs feature and it will have the same
look and feel with related issues widget.
However, related issues is a EE only feature but related MRs is not.
That’s why, to be able to reuse existing components,
I move them to ce/vue_shared and updated references.