Change content type detection header size to 4k
There is a problem detecting some file types like SVG. Depending
on how the file is built, it will detect the right content type
or not. For example, if the SVG file has a big html comment
(more than 512 bytes) at the beginning of it, when we start
the detection process we'll get the first 512 bytes and based on
that we will set the content type `text/plain`. But this is wrong
because the file is indeed an SVG.