Only exclude commit authors in merge request approvers
With https://gitlab.com/gitlab-org/gitlab-ee/merge_requests/9007, anyone
who rebases or amends a commit in a merge request will be tagged as a
committer, even if there were no changes. This is surprising behavior
and prevents reviewers from even touching a merge request.

Instead, we loosen the restriction to look only at the commit authors.

Closes https://gitlab.com/gitlab-org/gitlab-ee/issues/9787