Add DesignManagement::Version model
The DesignManagement::Version will keep track of different versions of
a design that were uploaded. Every time a new version gets uploaded we
create a new version so we know the last commit that touched the file.

It will also make it easy to track which version was commented on.