Fix SSH pull mirrors not working
This fixes a regression caused by
https://gitlab.com/gitlab-org/gitlab-ee/merge_requests/9134, which
causes an already-created SSH public/private key to be regenerated.
When the `Detect Host Keys` button is clicked, a Sidekiq worker
generates an SSH private and public key, the latter of which is shown to
the user. Creating a new mirror would wipe that data and cause a new key
to be generated unnecessarily.

The fix adds a hidden element so that the authentication mode will be
sent along the password to remove the need to clear the credentials
entirely.

Closes https://gitlab.com/gitlab-org/gitaly/issues/1550