Avoid scrollbar in more maintainable way
What this (and the previous) approach aims to achieve is *not* to show
a scrollbar in the severity filter dropdown. It's a fixed list of
severities, and inside a `dropdown-content`, the dropdown items are only
*just* tall enough to be taller than the max-height of
`dropdown-content`, making it render with a scrollbar, looking terrible.

The report type filter dropdown is also a fixed list of only four items,
so also doesn't need an internal scrollbar.

The projects filter dropdown, however, is not a fixed list, and requires
the internal scrollbar provided by `dropdown-content`.