Handle blank projects in CommitCollection#enrich!
A project is needed in order to fetch data from gitaly. Projects can be
absent from commits in certain rare situations (like when viewing a MR
of a deleted fork). In these cases, assume that the enriched data is not
needed.

See this comment:

https://gitlab.com/gitlab-org/gitlab-ce/merge_requests/26144#note_152191106

"It's led to a dilemma about where to "fix" this in code. I'm going
to fix it by allowing CommitCollection#enrich! to just return unenriched
commits when a project is missing, essentially "silently failing". I
hope this is the right decision. It's going with the assumption that
calls in these situations in the future are not needing the full data.
The alternative would be to allow CommitCollection#enrich! to still
error, but handle it in the methods that call #enrich!, however that
might lead to brittleness in future when working with project-less MRs."