Notify owner that group is invalid when LDAP "Sync now" fails
If a group is in an invalid state, it will never transition properly to
the right LDAP sync state, leading to confusion as to what the problem
is. In the previous implementation, clicking on the "Sync now" button
would erronenously tell the owner, "The group sync is already
scheduled", when in fact, it was stuck in the "ready" state.

This would help diagnose
https://gitlab.com/gitlab-org/gitlab-ee/issues/1529 and
https://gitlab.com/gitlab-org/gitlab-ee/issues/9613.