Render Key-value label differently
A label is a key-value label if it contains the following string `::`.
In case if a label is a key-value label, we'll change how a label
is displayed on the UI. We show an info icon next to the label name
which links to the documentation of how a key-value label is different
from a normal label. Rendering for old-style label remains the same.