[Geo] Don't mark sync as successful if repo does not exist
A secondary node marks every sync as successful if the repository
does not exist. While it's OK for the majority of the cases it
does not take into account some possible problems with
NFS servers availability.
To make more precise decisions we could make use of repository checksum.
If one exists but the repo doesn't, it's an indicator that
something is wrong with the infrastructure.