Pick reviewers based on branch name
Change reviewer roulette to always pick the same reviewers for the same
branch name. We do this by:

1. Making the branch name 'canonical' across CE and EE by stripping a
   leading 'ce-' or 'ee-' and a trailing '-ce' or '-ee'. If people are
   following our branch naming guidelines, this should give the same
   branch name in both repos.
2. Converting the branch name to a stable integer by taking the integer
   form of its MD5.
3. Passing that integer as a seed to Ruby's `Random` class, which 'may
   be used to ensure repeatable sequences of pseudo-random numbers
   between different runs of the program' (from the Ruby documentation).

The upshot is that the same branch name (in CE and EE) should always
pick the same reviewers, and those should be evenly distributed across
the set of possible reviewers due to the use of MD5.