Automatically enable multiple MR assignees feature flag
This schedules a background migration that checks whether
`multiple_merge_request_assignees` feature flag needs to
be enabled. The rule for it is that
every PopulateMergeRequestAssigneesTable background job
cannot be still running, or being retried or dead.

If there are any enqueued jobs in the state above, the
"checker" (MergeRequestAssigneesMigrationProgressCheck) will
be rescheduled for 3 hours ahead. If there are only dead jobs
in the queue (i.e. no scheduled/enqueued or retrying jobs), we
raise an error and stop the rescheduling.