Fix UrlBlocker spec
The newer version of HTTParty we now use will try harder to get a valid
URL out of a relative one. Unfortunately, when there's no room (as was
the case in this spec), the relative URL begins with `//`, which is
actually a protocol-relative URL, not a simple path.

Adding a room - which will always be the case in 'real life' - fixes
this.