Initial support for merge request blocks
This commit adds a database migration and ActiveRecord models to
represent the idea of a merge request that blocks another merge request
from being blocked.

There is no way to add a block through the UI yet, but you can create
them in the Rails console and observe that a merge request becomes
unmergeable as a result.