Backend work to comment on vulnerability dismissal
This change adds columns to the vulnerability_feedback table to support
adding a single plain-text comment for vulnerability feedback items and
specifically for commenting on dismissals of vulnerabilities.
It also adds support for submitting a comment when dismissing a
vulnerability as well as returning the comment-related attributes when
retrieving vulnerability_feedback records.