Fix an edge-case with a DATETIME column on MySQL
After a migration spec that would rollback the DB to a state prior to
ee/db/migrate/20170509153720_create_project_mirror_data_ee.rb, the
`project_mirror_data.next_execution_timestamp` column would not be
allowed to be NULL anymore, which contradicts with its definition from
`db/schema.rb` and would lead to further specs to fail.

See https://gitlab.com/gitlab-org/gitlab-ee/-/jobs/217401507 for an
example.

That was only discovered because of bad luck in this MR:
https://gitlab.com/gitlab-org/gitlab-ee/merge_requests/13149?commit_id=588165d03440b139e4241447dbeb8e368044abf6#note_173384953

[run mysql]

Signed-off-by: Rémy Coutable <remy@rymai.me>