Respect limited indexing settings in rake tasks
Some of the rake tasks were not respecting the new limited indexing
settings for Elasticsearch, so this refactors them to use
IndexRecordService through ElasticIndexerWorker. As part of this change
we're also queuing a Sidekiq job for each individual project now
instead of processing them in batches in each job, and we're always
processing them asynchronously.

- The index_repositories, index_repositories_async, index_database and
  index_$MODEL tasks were replaced with a single index_projects task,
  which indexes projects and all their associated records and their
  repositories

- The BATCH environment variable was removed because it's not useful
  anymore, since everything gets queued in Sidekiq anyway