Does not redirect user to the custom home page URL on a Geo secondary
If a user is not signed-in and tries to access root_path on a Geo
secondary node, redirects them to the sign-in page. Don't redirect
to the custom home page URL if one is present. Otherwise, it
will break the Geo OAuth workflow always redirecting the user to
the Geo primary node, which prevents access to the secondary node.