Add Dependency List alert components
This forms part of the [Dependency List MVC][1].

These alerts are shown during the edge cases documented in that issue.

This adds a new somewhat generic alert component that for the most part
matches the [GitLab design spec][2], but it's local to the dependency
list feature. This local implementation is expected to be replaced soon
with a generic implementation of alerts, implemented in
[`gitlab-ui`][3]. Once that's done, this feature-specific implementation
can be removed/swapped out in favour of that.

[1]: https://gitlab.com/gitlab-org/gitlab-ee/issues/10075
[2]: https://design.gitlab.com/components/alerts/
[3]: https://gitlab.com/gitlab-org/gitlab-ui/issues/210