Add `rule_type` to indexes explicitly
This means that we're enforcing uniqueness of a rule name per approval
merge request type on MySQL, since partial indexes are not supported
there.

On postgres, `rule_type` was already part of the index, but as the
index is partial it would not apply for rules that are not code owner
rules.

[run mysql]