Add trait to create orphaned artifact Geo registry
It looks complicated, but it's really powerful. With these factory
changes you're allowed to:

  create(:geo_job_artifact_registry, :orphan)

It will fill in `artifact_id` in the registry, to a non-existing
artifact.

And...

  create(:geo_job_artifact_registry, :orphan, artifact_type: :archive)

Will do the same, but it will also use the `ci_job_artifact` factor to
also create the file on disk. So you'll have an orphaned job artifact
file on disk, with a matching Geo job artifact registry.