Teach RolloutStatus about has_legacy_app_label
This will be used by the wrong end to display a warning about deploy
boards when the app label matching the environment slug is used.

has_legacy_app_label will be true whenever we detect a deployment using
the legacy label.

NB: If a deployment uses the app label but does not match the
environment slug, don't match it. This could happen eg when user deploys
a un-related helm chart to the same namespace. Adds a test case for this
scenario.